requirements.txt
pandas>=2.0.0
numpy>=1.24.0
faker>=18.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
fpdf2>=2.7.0
pydantic>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
nbformat>=5.9.0
Pillow>=10.0.0
nbconvert>=7.0.0
